requirements.txt
python-dotenv>=1.0.0
streamlit>=1.28.0
crewai>=0.86.0
crewai-tools>=0.17.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
